Split http server and http handler
